fix: BGE-325 block-scoped namespace in AchievementMapper + centralize GetForUser pipeline
- Convert AchievementMapper.cs to block-scoped namespace to match all other Controllers/ files
- Extract LINQ orchestration pipeline into AchievementMapper.GetForUser(GlobalState, string userId)
- Both PlayerManagementController.GetMyAchievements and PlayersController.GetAchievements
  now delegate to GetForUser; only the userId argument differs
